test(sync): re-split engine tests on top of the merged pull pipeline
The merge took master's monolithic engine_test.go, so this re-applies
the issue #736 split over the current content. Functions pack
sequentially into engine_test.go through engine_7_test.go, every file
stays under 1000 lines, and each file carries only the imports it uses.
No test body changed.
